backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
psycopg2-binary
google-generativeai
